Implementacion Customhook
